package/fastapi-mongo/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
authlib==1.3.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cryptography==42.0.7kw
deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mongo==0.0.5
fastapi-sessions==0.3.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
limits==3.12.0
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
motor==3.4.0
orjson==3.10.3
packaging==24.1
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.3
pydantic-core==2.18.4
pygments==2.18.0
pymongo==4.7.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pyyaml==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
